OK LETS GET WEIRD
I'm scaffolding out some places we can hang more ideas. The notion is
this:

A Doctopus is a Thing. It is an Organization of some kind -- the highest
leve logical group.

A Doctopus has one or more heads. Each head is a group of docs and the
orchestration for them.

Each Head controls some number of tentacles. Each tentacle is a source
of documentation and enough info to typeset it.